requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
decorator==5.2.1
Django==5.1.7
djangorestframework==3.15.2
drf-yasg==1.21.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
lxml==5.4.0
moviepy==2.1.2
numpy==2.2.5
packaging==24.2
pdf2image==1.17.0
pillow==10.4.0
proglog==0.1.11
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
XlsxWriter==3.2.3
